Entities and access update

+++ Entities/Concrete/RoomMember.cs
@@ -7,7 +7,7 @@ namespace Entities.Concrete
         public int RoomMemberId { get; set; }
         public int RoomId { get; set; }
         public int UserId { get; set; }
-        public int RankId { get; set; }
+        public byte RankId { get; set; }
     }
 
 }
